Rename create_session.py to session.py
in anticipation of it having more functions
